Move molluscoid to separate file
