+++ src/common.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <vector>
+
+#include <opencv/cv.h>
+
+using contour = std::vector<cv::Point2i>;
+
+
