+++ src/detect.h
@@ -6,23 +6,8 @@
 
 #include <opencv/cv.h>
 
-using contour = std::vector<cv::Point2i>;
-
-std::vector<cv::Mat> process(std::string imageName, std::string inDir);
-std::string getColor(cv::Mat image);
-
-struct molluscoid {
-	const contour cont;
-	const cv::Mat image;
-	const cv::RotatedRect bounding_rect;
-
-	std::string get_color() const;
-	double ratio() const;
-	double angle() const;
-	
-	molluscoid(const contour& _cont, const cv::Mat& _image);
-	~molluscoid();
-};
+#include "molluscoid.h"
+#include "common.h"
 
 class moldec {
 	const cv::Mat origin;
@@ -45,3 +30,4 @@ public:
 
 	void write_images(std::ostream& os, const std::vector<std::string>& data, const std::string& imageName) const;
 };
+
